AGENTNEW/requirements.txt
streamlit==1.39.0
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
yake==0.4.8
nltk==3.9.1
rank-bm25==0.2.2
rapidfuzz==3.9.6
sentence-transformers==3.0.1
scikit-learn>=1.5.0
pymupdf==1.24.10
pdfminer.six==20240706
beautifulsoup4==4.12.3
lxml==5.3.0
networkx==3.3
sumy==0.11.0
numpy>=1.26.0
pandas>=2.2.2
